requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
lxml==4.9.0
mysql-connector-python==8.0.29
protobuf==4.21.1
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
